✨ feat: user type auth.ts로 이동
